Add input mode flags and update API accessors
• Added new HudInputMode flags to indicate whether text input is enabled or if only cursor input is enabled
• Updated API accessors to support this and changes to TextBoard accessors made to support clipping

+++ RichHudMaster/Data/Scripts/RichHudFramework/Shared/UI/HUD/HudElementInterfaces/IReadOnlyHudParent.cs
@@ -18,7 +18,7 @@ namespace RichHudFramework
         >;
 
         [Flags]
-        public enum HudElementStates : uint
+        public enum HudElementStates : ushort
         {
             None = 0x0,
             IsVisible = 0x1,
